Add index.html and index.js
I've lost track of time to commit every changes in my code.
But for now the rock, paper, and scissors game is done.

I plan to add some features like waiting time before prompting the user
again for a choice.
Also, anyone who reach first 5(scores) should be the winner.
For now it works, because the game is limited to be only played 5
rounds per game.