Modify logic and design
As for the UI aspect, I've changed the font-family, added some icons for
the buttons(rock, paper and scissor).

Added a heading for the scores of human and computer

As for the logic, I am still trying to understand the use of
`textContent` and how helpful it is.

I also changed my eventListener for each button and used instead a
forEach approach. And from there, the selections are stored and passed to
`playRound()`.