Add `getWinner()` to determine the winner of the game.
Added a new section where the winner will be announced. Also, minimized the size of the scores.

Removed the `getHumanChoice()` and `playGame()`. the `ROUNDS` is also removed.
Overall, the game is working as how it is to work.

The only problem is that when a player clicks in the center part of the button, the one gets clicked
is the `<img>`. And, as for the `getChoiceName()` it returns "Unknown".

The rounds is removed because the logic is that whoever reaches 5 points first will be the winner.